HSEARCH-3619 Move the complexity of automatic indexing from HibernateSearchContextService to HibernateOrmSearchSession
That's a first step to remove some of the cyclic dependencies.
