+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/spi/SearchSessionImplementor.java
@@ -7,8 +7,6 @@
 package org.hibernate.search.mapper.orm.session.spi;
 
 import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
-import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
-import org.hibernate.search.mapper.orm.session.AutomaticIndexingSynchronizationStrategy;
 import org.hibernate.search.mapper.orm.session.SearchSession;
 import org.hibernate.search.mapper.pojo.work.spi.PojoSessionWorkExecutor;
 import org.hibernate.search.mapper.pojo.work.spi.PojoWorkPlan;
@@ -19,16 +17,10 @@ public interface SearchSessionImplementor extends AutoCloseable, SearchSession {
 	void close();
 
 	/**
-	 * @param commitStrategy The commit strategy to use for every affected index.
-	 * @param refreshStrategy The refresh strategy to use for every affected index.
-	 * @return A new work plan for this session, maintaining its state (list of works) independently from the session.
-	 * Calling {@link PojoWorkPlan#execute()} is required to actually execute works,
-	 * the session will <strong>not</strong> do it automatically upon closing.
+	 * @return The {@link PojoWorkPlan} to use for changes to entities in the given session.
 	 */
-	PojoWorkPlan createWorkPlan(DocumentCommitStrategy commitStrategy, DocumentRefreshStrategy refreshStrategy);
+	PojoWorkPlan getCurrentWorkPlan();
 
 	PojoSessionWorkExecutor createSessionWorkExecutor(DocumentCommitStrategy commitStrategy);
-
-	AutomaticIndexingSynchronizationStrategy getAutomaticIndexingSynchronizationStrategy();
 
 }
